+++ cmd/manual.go
@@ -67,7 +67,17 @@ var manualCmd = &cobra.Command{
 
 		format, _ := cmd.Flags().GetString("format")
 		asJSON, _ := cmd.Flags().GetBool("json")
-		return newEntry(c, tei, viper.GetBool(INTERACTIVE), viper.GetBool(ALLOW_PROJECT_NAME), false, format, asJSON)
+		return manageEntry(
+			c,
+			tei,
+			createTimeEntry(c),
+			viper.GetBool(INTERACTIVE),
+			viper.GetBool(ALLOW_PROJECT_NAME),
+			false,
+			format,
+			asJSON,
+			true,
+		)
 	}),
 }